Change: re-construct files
