+++ src/Commands/index.ts
@@ -17,42 +17,9 @@
  */
 
 import type { CommandMessage } from "../Component/CommandMessage";
-import type { PageToggle } from "../Component/PageToggle";
-import type { GuildDataContainer } from "../Structure";
-import type { MusicBot } from "../bot";
-import type { categories } from "./commands";
-import type { Client } from "eris";
+import type { ListCommandInitializeOptions, UnlistCommandInitializeOptions, ListCommandWithArgumentsInitializeOptions, CommandArgs, SlashCommandArgument } from "../Structure/Command";
 
-import * as fs from "fs";
-import * as path from "path";
-
-import { LogEmitter } from "../Structure";
-
-type BaseCommandInitializeOptions = {
-  name:string,
-  alias:Readonly<string[]>,
-};
-
-type ListCommandWithArgumentsInitializeOptions = BaseCommandInitializeOptions & {
-  description:string,
-  unlist:boolean,
-  examples:string,
-  usage:string,
-  category:keyof typeof categories,
-  argument:SlashCommandArgument[],
-};
-
-type ListCommandWithoutArgumentsInitializeOptions = BaseCommandInitializeOptions & {
-  description:string,
-  unlist:false,
-  category:keyof typeof categories,
-};
-
-type ListCommandInitializeOptions = ListCommandWithArgumentsInitializeOptions | ListCommandWithoutArgumentsInitializeOptions;
-
-type UnlistCommandInitializeOptions = BaseCommandInitializeOptions & {
-  unlist:true,
-};
+export { CommandArgs } from "../Structure/Command";
 
 /**
  * すべてのコマンドハンドラーの基底クラスです
@@ -88,117 +55,5 @@ export abstract class BaseCommand {
       this._category = category;
       this._argument = argument || null;
     }
-  }
-}
-
-/**
- * スラッシュコマンドの引数として取れるものを定義するインターフェースです
- */
-export interface SlashCommandArgument {
-  type:"bool"|"integer"|"string";
-  name:string;
-  description:string;
-  required:boolean;
-  choices?:{[key:string]:string|number};
-}
-
-/**
- * コマンドのランナに渡される引数
- */
-export interface CommandArgs {
-  /**
-   * ボットのインスタンス
-   */
-  bot:MusicBot;
-  /**
-   * ボットのサーバーデータ
-   */
-  server:GuildDataContainer;
-  /**
-   * コマンドの生の引数
-   */
-  rawArgs: string;
-  /**
-   * コマンドのパース済み引数
-   */
-  args: string[];
-  /**
-   * 生存しているPageToggleの配列
-   */
-  embedPageToggle:PageToggle[];
-  /**
-   * ボットのクライアント
-   */
-  client:Client;
-  /**
-   * サーバーデータの初期化関数
-   * @param guildid サーバーID
-   * @param channelid チャンネルID
-   */
-  initData: (guildid:string, channelid:string) => void;
-}
-
-/**
- * コマンドマネージャー
- */
-export class CommandsManager extends LogEmitter {
-  private static _instance = null as CommandsManager;
-  /**
-   * コマンドマネージャーの唯一のインスタンスを返します
-   */
-  static get Instance(){
-    if(this._instance) return this._instance;
-    else return this._instance = new CommandsManager();
-  }
-
-  /**
-   * コマンドマネージャーの唯一のインスタンスを返します
-   */
-  get Commands(){
-    return this.commands;
-  }
-
-  private readonly commands = null as BaseCommand[];
-
-  private constructor(){
-    super();
-    this.setTag("CommandsManager");
-    this.Log("Initializing");
-    this.commands = [];
-    fs.readdirSync(__dirname, {withFileTypes: true})
-      .filter(d => d.isFile())
-      .map(d => d.name)
-      .filter(n => n.endsWith(".js") && n !== "index.js")
-      .map(n => n.slice(0, -3))
-      .forEach(n => {
-        // eslint-disable-next-line new-cap
-        const cp = new (require(path.join(__dirname, n)).default)() as BaseCommand;
-        this.commands.push(cp);
-        return cp;
-      });
-    this.Log("Initialized");
-  }
-
-  /**
-   * コマンド名でコマンドを解決します
-   * @param command コマンド名
-   * @returns 解決されたコマンド
-   */
-  resolve(command:string){
-    this.Log("Resolving command");
-    let result = null;
-    for(let i = 0; i < this.commands.length; i++){
-      if(this.commands[i].name === command || this.commands[i].alias.includes(command)){
-        result = this.commands[i];
-        break;
-      }
-    }
-    if(result) this.Log(`Command "${command}" was resolved successfully`);
-    else this.Log("Command not found");
-    return result;
-  }
-
-  Check(){
-    return true;
   }
 }
