+++ src/Structure/index.ts
@@ -18,6 +18,7 @@
 
 export * from "./LogEmitter";
 export * from "./ManagerBase";
+export * from "./Command";
 export * from "./GuildDataContainer";
 export * from "./SearchPanel";
 export * from "./YmxFormat";
